+++ FileSuffixesTest/FileSuffixesTest/AppDelegate.h
@@ -1,0 +1,15 @@
+//
+//  AppDelegate.h
+//  FileSuffixesTest
+//
+//  Created by Nick Lockwood on 09/06/2012.
+//  Copyright (c) 2012 Charcoal Design. All rights reserved.
+//
+
+#import <Cocoa/Cocoa.h>
+
+@interface AppDelegate : NSObject <NSApplicationDelegate>
+
+@property (assign) IBOutlet NSWindow *window;
+
+@end