+++ FileSuffixesTest/FileSuffixesTest/AppDelegate.m
@@ -1,0 +1,18 @@
+//
+//  AppDelegate.m
+//  FileSuffixesTest
+//
+//  Created by Nick Lockwood on 09/06/2012.
+//  Copyright (c) 2012 Charcoal Design. All rights reserved.
+//
+
+#import "AppDelegate.h"
+
+@implementation AppDelegate
+
+- (void)applicationDidFinishLaunching:(NSNotification *)aNotification
+{
+    // Insert code here to initialize your application
+}
+
+@end