+++ Tests/FileSuffixesTest/MacClasses/AppDelegate.m
@@ -1,0 +1,21 @@
+//
+//  AppDelegate.m
+//  FileSuffixesTestMac
+//
+//  Created by Nick Lockwood on 09/06/2012.
+//  Copyright (c) 2012 Charcoal Design. All rights reserved.
+//
+
+#import "AppDelegate.h"
+
+@implementation AppDelegate
+
+@synthesize window;
+@synthesize windowController;
+
+- (void)applicationDidFinishLaunching:(NSNotification *)aNotification
+{
+    // Insert code here to initialize your application
+}
+
+@end