+++ Tests/FileSuffixesTest/MacClasses/WindowController.h
@@ -1,0 +1,17 @@
+//
+//  WindowController.h
+//  FileSuffixesTest
+//
+//  Created by Nick Lockwood on 09/06/2012.
+//  Copyright (c) 2012 Charcoal Design. All rights reserved.
+//
+
+#import <Cocoa/Cocoa.h>
+
+
+@interface WindowController : NSWindowController
+
+@property (nonatomic, strong) IBOutlet NSImageView *imageView1;
+@property (nonatomic, strong) IBOutlet NSImageView *imageView2;
+
+@end